social-lead-agent/requirements.txt
langchain
langgraph
langchain-groq
faiss-cpu
sentence-transformers
streamlit
python-dotenv